+++ game_service/client/vault_client.py
